Refactored but not tested (jar file required)
