+++ JavaBackup/AllProjects/src/tools/ObserverSimple.java
@@ -1,0 +1,5 @@
+package tools;
+
+public interface ObserverSimple<T> {
+	public void notify(T something);
+}
