+++ JavaBackup/AllProjects/src/tools/remoteControl/controllerModel/ConnectionStatus.java
@@ -1,0 +1,7 @@
+package tools.remoteControl.controllerModel;
+
+import java.io.Serializable;
+
+public enum ConnectionStatus implements Serializable {
+	NewBorn, Connecting, Connected, ShuttedDown;
+}
