+++ JavaBackup/AllProjects/src/tools/remoteControl/msg/MessageTypeRC.java
@@ -1,0 +1,7 @@
+package tools.remoteControl.msg;
+
+import java.io.Serializable;
+
+public enum MessageTypeRC implements Serializable {
+	Ping, ShutDown, Mouse, Keyboard, Image
+}
