+++ JavaBackup/AllProjects/src/tools/remoteControl/msg/MouseAction.java
@@ -1,0 +1,7 @@
+package tools.remoteControl.msg;
+
+import java.io.Serializable;
+
+public enum MouseAction implements Serializable {
+	Move, Press, Release, Wheel, Click
+}
